Fix overflow
The calibration variables _cal_vi1 and _cal_vj1 were uint16_t, but this led to overflow on my screen, making the
touch x-position jump from around 100 to 56,000 in the lower 25% of the screen. Changing these variables to int32_t
fixes this overflow.

+++ XPT2046.h
@@ -53,7 +53,7 @@ private:
   uint8_t _cs_pin, _irq_pin;
 
   int32_t _cal_dx, _cal_dy, _cal_dvi, _cal_dvj;
-  uint16_t _cal_vi1, _cal_vj1;
+  int32_t _cal_vi1, _cal_vj1;
 
   uint16_t _readLoop(uint8_t ctrl, uint8_t max_samples) const;
 };
